requirements.txt
androguard==3.3.5
Appium-Python-Client==2.11.1
asn1crypto==1.5.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
future==0.18.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
protobuf==4.21.12
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
pytest-html==4.0.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.11.2
selenium-page-factory==2.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
wsproto==1.2.0
allure-pytest
